terminado

+++ routes/teams.js
@@ -10,7 +10,7 @@ export const router = express.Router();
 
 router.get('/', async (req, res) => {
     try{
-        const teams = await Team.find();
+        const teams = await Team.find().populate('roster');
 
         if(teams.length <= 0) {
             const error = Error()
@@ -18,11 +18,11 @@ router.get('/', async (req, res) => {
             throw error;
         }
 
-        res.status(200).send(teams)
+        res.status(200).send({result: teams})
 
     } catch(ex) {
         const error = getErrorMessage(ex);
-        res.status(error.statusCode).send(error.message);
+        res.status(error.statusCode).send({error: error.message});
     }
 });
 
@@ -31,7 +31,7 @@ router.get('/:id', async (req, res) => {
     {
         const { id } = req.params;
         
-        const team = await Team.findById(id);
+        const team = await Team.findById(id).populate('roster');
 
         if(!team) {
             const err = new Error()
@@ -39,77 +39,87 @@ router.get('/:id', async (req, res) => {
             throw err;
         }
 
-        const rostersOfTeam = await Roster.find({_id: {$in: team.roster}})
+        await team.populate('roster.player')
 
-        team.roster = rostersOfTeam.filter(r => r.active);
-
-        res.status(200).send(team);
+        team.roster = team.roster.filter(r => r.active);
+        res.status(200).send({result: team});
     }
     catch(ex){
         console.log(ex.cause)
         const error = getErrorMessage(ex);
-        res.status(error.statusCode).send(error.message)
+        res.status(error.statusCode).send({error: error.message})
     }
 })  
 
 router.post('/', async (req, res) => {
     try {
-        const team = new Team({... body});
+        const team = new Team({... req.body});
         const savedTeam = await team.save();
 
-        res.status(201).send(savedTeam);
+        res.status(201).send({result: savedTeam});
     } catch(ex) {
         const error = getErrorMessage(ex);
-        res.status(error.statusCode).send(error.message)
+        res.status(error.statusCode).send({error: error.message})
     }
 })
 
 router.post('/:id/roster', async (req, res) => {
     try{
-        const roster = new Roster({...req.body});
         const teamId = req.params.id;
-        const playerId = roster.player;
+        const roster = new Roster({...req.body});
 
-        const team = await Team.findById(teamId);
-        const player = await Player.findById(playerId);
-
-        if(!team){
-            const error = new Error("el equipo no se pudo encontrar o no existe");
-            error.name = "NotFound"
-            throw error;
-        }
+        const team = await Team.findById(teamId)
+        console.log("Equipo encontrado: ", team)
         
-        if(!player){
-            const error = new Error("el jugador no se pudo encontrar o no existe");
-            error.name = "NotFound"
-            throw error;
+        if(!team) {
+            const err = new Error();
+            err.name = "TeamNotFound";
+            throw err;
         }
 
-        const roastersWithPlayer = await Roster.find({player: playerId, active: true})
-        const teamsWithActiveRoster = await Team.find({roster: {$in: roastersWithPlayer}})
+        const roastersWithPlayer = await Roster.find({player: roster.player})
+        const rostersWithActivePlayer = roastersWithPlayer.filter(r => r.active)
 
-        if(teamsWithActiveRoster.filter(t => t != team).length > 0)
-        {
+        const teamWithActiveRoster = await Team.findOne({roster: {$in: rostersWithActivePlayer}});
+        const teamWithRoster = await Team.findOne({roster: {$in: rostersWithActivePlayer}});
+
+        const rosterInThisTeam = await Roster.findOne({_id: {$in: team.roster}, player: roster.player});
+
+        if(teamWithActiveRoster && teamWithActiveRoster._id.equals(team._id)) {
             const error = new Error();
             error.name = "AlreadyActiveInThisTeam"
             throw error;
         }
 
-        if(teamsWithActiveRoster.length > 0)
-        {
+        if(teamWithActiveRoster) {
             const error = new Error();
             error.name = "AlreadyActive"
             throw error;
         }
 
-        const savedRoster = await roster.save();
-        team.roster.push(savedRoster);
-        const savedTeam = await Team.findByIdAndUpdate(team._id, team);
+        let savedTeam;
 
-        res.status(200).send(team);
+        if(rosterInThisTeam) {
+            rosterInThisTeam.active = true;
+            const savedRoster = await rosterInThisTeam.save();
+            savedTeam = await team.save();
+        }
+        else {
+            const savedRoster = await roster.save();
+            team.roster.push(savedRoster);
+            savedTeam = await team.save();
+        }
+
+        await savedTeam.populate('roster');
+        await savedTeam.populate('roster.player')
+        
+        savedTeam.roster = savedTeam.roster.filter(r => r.active)
+
+        res.status(200).send({result: savedTeam});
+
     }catch(ex){
         const error = getErrorMessage(ex);
-        res.status(error.statusCode).send(error.message)
+        res.status(error.statusCode).send({error: error.message})
     }
 });
 
@@ -145,20 +155,18 @@ router.delete('/:id/roster/:playerId', async (req, res) => {
         const savedTeam = Team.findByIdAndUpdate(teamId, team)
         console.log(activePlayer);
 
-        res.status(200).send(team);
+        res.status(200).send({result: team});
 
     }catch(ex){
         const error = getErrorMessage(ex);
-        res.status(error.statusCode).send(error.message)
+        res.status(error.statusCode).send({error: error.message})
     }
 })
 
 router.delete('/:id', async (req, res) => {
     try{
         const teamId = req.params.id;
-
         const team = await Team.findById(teamId);
-
         if(!team) {
             const error = new Error();
             error.name = "TeamNotFound"
@@ -177,13 +185,25 @@ router.delete('/:id', async (req, res) => {
             throw error;
         }
 
+        for (const r of team.roster) {
+            await Roster.findByIdAndDelete(r);
+        }
+
+
+        if(team.roster > 0) {
+            const error = Error()
+            error.name = "CannotDeleteRoster"
+            throw error;
+        }
+
         await team.deleteOne();
 
-        res.status(200).send(team);
+        res.status(200).send({result: team});
+
 
     }catch(ex){
         const error = getErrorMessage(ex);
-        res.status(error.statusCode).send(error.message)
+        res.status(error.statusCode).send({error: error.message})
     }
 });
 
@@ -192,9 +212,12 @@ function getErrorMessage(er) {
 
     if(er.name == 'EmptyList') { return { statusCode: 404, message: 'No existen equipos registrados en el sistema' } }
     if(er.name == 'TeamNotFound') { return { statusCode: 404, message: 'No existe ese equipo en el sistema.' } }
-    if(er.name == 'AlreadyActive') { return { statusCode: 400, message: 'No se puede eliminar el jugador porque está activo en algún equipo' } }
-    if(er.name == 'AlreadyActiveInThisTeam') { return { statusCode: 400, message: 'No se puede eliminar el jugador porque está activo en este equipo' } }
+    if(er.name == 'PlayerNotFound') { return { statusCode: 404, message: 'Jugador no encontrado.' } }
+    if(er.name == 'AlreadyActive') { return { statusCode: 400, message: 'El jugador está activo en otro equipo.' } }
+    if(er.name == 'AlreadyActiveInThisTeam') { return { statusCode: 400, message: 'El jugador ya está activo en el roster de este equipo.' } }
     if(er.name == 'PlayerNotActive') { return { statusCode: 404, message: 'Jugador no encontrado.' } }
+    if(er.name == 'CannotDeleteRoster') { return { statusCode: 400, message: 'No se puede eliminar los rosters del equipo' } }
+    if(er.name == 'HasMatches') { return { statusCode: 400, message: 'No se puede eliminar el equipo porque tiene partidos asociados.' } }
 
     if(er.name == 'NeedParameters') { return { statusCode: 400, message: 'Falta el parámetro de búsqueda.' } }
     if(er.name == 'ValidationError') { return { statusCode: 400, message: 'Datos incorrectos: faltan campos obligatorios' } }
